+++ lib/rubocop/cop/rspec/message_spies.rb
@@ -36,7 +36,7 @@ module RuboCop
         SUPPORTED_STYLES = %w(have_received receive).freeze
 
         def_node_matcher :message_expectation, %(
-          (send (send nil :expect (send nil $_)) :to ...)
+          (send (send nil :expect $(...)) {:to :to_not :not_to} ...)
         )
 
         def_node_search :receive_message, %(
@@ -70,7 +70,7 @@ module RuboCop
           when :receive
             MSG_RECEIVE
           when :have_received
-            MSG_HAVE_RECEIVED % receiver
+            MSG_HAVE_RECEIVED % receiver.source
           end
         end
       end